# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Beauty Booking API
  version: 1.0.0
  description: >
    API для запису клієнтів до майстрів, керування послугами/розкладом, замовленнями продуктів/напоїв та відгуками.
servers:
  - url: http://localhost:8080
    description: Local Spring Boot
...
security:
  - bearerAuth: []
paths:
...
  /schedules:
    get:
      tags: [Schedules]
      summary: List schedules (time slots)
      parameters:
        - in: query
          name: masterId
          schema: { type: integer, format: int64 }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: available
          schema: { type: boolean }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SchedulesPage" }
    post:
      tags: [Schedules]
      summary: Create schedule slot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "400":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      in: query
      name: offset
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Invalid request }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string, example: email }
              issue: { type: string, example: must be a valid email }
      required: [code, message]
...
    Schedule:
      type: object
      additionalProperties: false
      properties:
        id: { type: integer, format: int64 }
        masterId: { type: integer, format: int64 }
        workDate: { type: string, format: date }
        startTime: { type: string, example: "10:00:00" }
        endTime: { type: string, example: "11:00:00" }
        available: { type: boolean }
      required: [id, masterId, workDate, startTime, endTime, available]
    ScheduleCreate:
      type: object
      additionalProperties: false
      properties:
        masterId: { type: integer, format: int64 }
        workDate: { type: string, format: date }
        startTime: { type: string, example: "10:00:00" }
        endTime: { type: string, example: "11:00:00" }
        available: { type: boolean, default: true }
      required: [masterId, workDate, startTime, endTime]
    SchedulesPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Schedule" }
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
      required: [items, limit, offset, total]
...